break: move `STATUS_CODES` to `response` module
