--- types/status.ts
@@ -1,5 +1,0 @@
-import { STATUS_CODES } from 'worktop';
-
-assert<string>(STATUS_CODES[200]);
-assert<string>(STATUS_CODES['200']);
-STATUS_CODES['404'] = 'Custom Error Message';
